+++ src/lib/types_test.rs
@@ -413,6 +413,7 @@ fn install_crate_info_eq_same_all() {
         min_version: Some("1.0.0".to_string()),
         version: Some("2.0.0".to_string()),
         install_command: Some("install".to_string()),
+        force: Some(true),
     };
     let second = InstallCrateInfo {
         crate_name: "test".to_string(),
@@ -424,6 +425,7 @@ fn install_crate_info_eq_same_all() {
         min_version: Some("1.0.0".to_string()),
         version: Some("2.0.0".to_string()),
         install_command: Some("install".to_string()),
+        force: Some(true),
     };
 
     assert_eq!(first, second);
@@ -441,6 +443,7 @@ fn install_crate_info_eq_same_no_component() {
         min_version: Some("1.0.0".to_string()),
         version: Some("2.0.0".to_string()),
         install_command: Some("install".to_string()),
+        force: Some(true),
     };
     let second = InstallCrateInfo {
         crate_name: "test".to_string(),
@@ -452,6 +455,7 @@ fn install_crate_info_eq_same_no_component() {
         min_version: Some("1.0.0".to_string()),
         version: Some("2.0.0".to_string()),
         install_command: Some("install".to_string()),
+        force: Some(true),
     };
 
     assert_eq!(first, second);
@@ -469,6 +473,7 @@ fn install_crate_info_eq_same_no_min_version() {
         min_version: None,
         version: Some("2.0.0".to_string()),
         install_command: Some("install".to_string()),
+        force: Some(true),
     };
     let second = InstallCrateInfo {
         crate_name: "test".to_string(),
@@ -480,6 +485,7 @@ fn install_crate_info_eq_same_no_min_version() {
         min_version: None,
         version: Some("2.0.0".to_string()),
         install_command: Some("install".to_string()),
+        force: Some(true),
     };
 
     assert_eq!(first, second);
@@ -497,6 +503,7 @@ fn install_crate_info_eq_same_no_version() {
         min_version: Some("1.0.0".to_string()),
         version: None,
         install_command: Some("install".to_string()),
+        force: Some(true),
     };
     let second = InstallCrateInfo {
         crate_name: "test".to_string(),
@@ -508,6 +515,7 @@ fn install_crate_info_eq_same_no_version() {
         min_version: Some("1.0.0".to_string()),
         version: None,
         install_command: Some("install".to_string()),
+        force: Some(true),
     };
 
     assert_eq!(first, second);
@@ -525,6 +533,7 @@ fn install_crate_info_eq_same_no_install_command() {
         min_version: Some("1.0.0".to_string()),
         version: Some("2.0.0".to_string()),
         install_command: None,
+        force: Some(true),
     };
     let second = InstallCrateInfo {
         crate_name: "test".to_string(),
@@ -536,6 +545,37 @@ fn install_crate_info_eq_same_no_install_command() {
         min_version: Some("1.0.0".to_string()),
         version: Some("2.0.0".to_string()),
         install_command: None,
+        force: Some(true),
+    };
+
+    assert_eq!(first, second);
+}
+
+#[test]
+fn install_crate_info_eq_same_no_force() {
+    let first = InstallCrateInfo {
+        crate_name: "test".to_string(),
+        binary: "bin".to_string(),
+        test_arg: TestArg {
+            inner: vec!["--help".to_string()],
+        },
+        rustup_component_name: Some("component".to_string()),
+        min_version: Some("1.0.0".to_string()),
+        version: Some("2.0.0".to_string()),
+        install_command: Some("install".to_string()),
+        force: None,
+    };
+    let second = InstallCrateInfo {
+        crate_name: "test".to_string(),
+        binary: "bin".to_string(),
+        test_arg: TestArg {
+            inner: vec!["--help".to_string()],
+        },
+        rustup_component_name: Some("component".to_string()),
+        min_version: Some("1.0.0".to_string()),
+        version: Some("2.0.0".to_string()),
+        install_command: Some("install".to_string()),
+        force: None,
     };
 
     assert_eq!(first, second);
@@ -553,6 +593,7 @@ fn install_crate_info_eq_different_crate_name() {
         min_version: Some("1.0.0".to_string()),
         version: Some("2.0.0".to_string()),
         install_command: Some("install".to_string()),
+        force: None,
     };
     let second = InstallCrateInfo {
         crate_name: "test2".to_string(),
@@ -564,6 +605,7 @@ fn install_crate_info_eq_different_crate_name() {
         min_version: Some("1.0.0".to_string()),
         version: Some("2.0.0".to_string()),
         install_command: Some("install".to_string()),
+        force: None,
     };
 
     assert!(first != second);
@@ -581,6 +623,7 @@ fn install_crate_info_eq_different_binary() {
         min_version: Some("1.0.0".to_string()),
         version: Some("2.0.0".to_string()),
         install_command: Some("install".to_string()),
+        force: None,
     };
     let second = InstallCrateInfo {
         crate_name: "test".to_string(),
@@ -592,6 +635,7 @@ fn install_crate_info_eq_different_binary() {
         min_version: Some("1.0.0".to_string()),
         version: Some("2.0.0".to_string()),
         install_command: Some("install".to_string()),
+        force: None,
     };
 
     assert!(first != second);
@@ -609,6 +653,7 @@ fn install_crate_info_eq_different_test_arg() {
         min_version: Some("1.0.0".to_string()),
         version: Some("2.0.0".to_string()),
         install_command: Some("install".to_string()),
+        force: None,
     };
     let second = InstallCrateInfo {
         crate_name: "test".to_string(),
@@ -620,6 +665,7 @@ fn install_crate_info_eq_different_test_arg() {
         min_version: Some("1.0.0".to_string()),
         version: Some("2.0.0".to_string()),
         install_command: Some("install".to_string()),
+        force: None,
     };
 
     assert!(first != second);
@@ -637,6 +683,7 @@ fn install_crate_info_eq_different_component_type() {
         min_version: Some("1.0.0".to_string()),
         version: Some("2.0.0".to_string()),
         install_command: Some("install".to_string()),
+        force: None,
     };
     let second = InstallCrateInfo {
         crate_name: "test".to_string(),
@@ -648,6 +695,7 @@ fn install_crate_info_eq_different_component_type() {
         min_version: Some("1.0.0".to_string()),
         version: Some("2.0.0".to_string()),
         install_command: Some("install".to_string()),
+        force: None,
     };
 
     assert!(first != second);
@@ -665,6 +713,7 @@ fn install_crate_info_eq_different_component_value() {
         min_version: Some("1.0.0".to_string()),
         version: Some("2.0.0".to_string()),
         install_command: Some("install".to_string()),
+        force: None,
     };
     let second = InstallCrateInfo {
         crate_name: "test".to_string(),
@@ -676,6 +725,7 @@ fn install_crate_info_eq_different_component_value() {
         min_version: Some("1.0.0".to_string()),
         version: Some("2.0.0".to_string()),
         install_command: Some("install".to_string()),
+        force: None,
     };
 
     assert!(first != second);
@@ -693,6 +743,7 @@ fn install_crate_info_eq_different_min_version_type() {
         min_version: Some("1.0.0".to_string()),
         version: Some("2.0.0".to_string()),
         install_command: Some("install".to_string()),
+        force: None,
     };
     let second = InstallCrateInfo {
         crate_name: "test".to_string(),
@@ -704,6 +755,7 @@ fn install_crate_info_eq_different_min_version_type() {
         min_version: None,
         version: Some("2.0.0".to_string()),
         install_command: Some("install".to_string()),
+        force: None,
     };
 
     assert!(first != second);
@@ -721,6 +773,7 @@ fn install_crate_info_eq_different_min_version_value() {
         min_version: Some("1.0.0".to_string()),
         version: Some("2.0.0".to_string()),
         install_command: Some("install".to_string()),
+        force: None,
     };
     let second = InstallCrateInfo {
         crate_name: "test".to_string(),
@@ -732,6 +785,7 @@ fn install_crate_info_eq_different_min_version_value() {
         min_version: Some("2.0.0".to_string()),
         version: Some("2.0.0".to_string()),
         install_command: Some("install".to_string()),
+        force: None,
     };
 
     assert!(first != second);
@@ -749,6 +803,7 @@ fn install_crate_info_eq_different_version_type() {
         min_version: Some("1.0.0".to_string()),
         version: Some("2.0.0".to_string()),
         install_command: Some("install".to_string()),
+        force: None,
     };
     let second = InstallCrateInfo {
         crate_name: "test".to_string(),
@@ -760,6 +815,7 @@ fn install_crate_info_eq_different_version_type() {
         min_version: Some("1.0.0".to_string()),
         version: None,
         install_command: Some("install".to_string()),
+        force: None,
     };
 
     assert!(first != second);
@@ -777,6 +833,7 @@ fn install_crate_info_eq_different_version_value() {
         min_version: Some("1.0.0".to_string()),
         version: Some("2.0.0".to_string()),
         install_command: Some("install".to_string()),
+        force: None,
     };
     let second = InstallCrateInfo {
         crate_name: "test".to_string(),
@@ -788,6 +845,7 @@ fn install_crate_info_eq_different_version_value() {
         min_version: Some("1.0.0".to_string()),
         version: Some("1.0.0".to_string()),
         install_command: Some("install".to_string()),
+        force: None,
     };
 
     assert!(first != second);
@@ -805,6 +863,7 @@ fn install_crate_info_eq_different_install_command_type() {
         min_version: Some("1.0.0".to_string()),
         version: Some("2.0.0".to_string()),
         install_command: Some("install".to_string()),
+        force: None,
     };
     let second = InstallCrateInfo {
         crate_name: "test".to_string(),
@@ -816,6 +875,7 @@ fn install_crate_info_eq_different_install_command_type() {
         min_version: Some("1.0.0".to_string()),
         version: Some("2.0.0".to_string()),
         install_command: None,
+        force: None,
     };
 
     assert!(first != second);
@@ -833,6 +893,7 @@ fn install_crate_info_eq_different_install_command_value() {
         min_version: Some("1.0.0".to_string()),
         version: Some("2.0.0".to_string()),
         install_command: Some("install1".to_string()),
+        force: None,
     };
     let second = InstallCrateInfo {
         crate_name: "test".to_string(),
@@ -844,6 +905,67 @@ fn install_crate_info_eq_different_install_command_value() {
         min_version: Some("1.0.0".to_string()),
         version: Some("2.0.0".to_string()),
         install_command: Some("install2".to_string()),
+        force: None,
+    };
+
+    assert!(first != second);
+}
+
+#[test]
+fn install_crate_info_eq_different_force_type() {
+    let first = InstallCrateInfo {
+        crate_name: "test".to_string(),
+        binary: "bin".to_string(),
+        test_arg: TestArg {
+            inner: vec!["--help".to_string()],
+        },
+        rustup_component_name: Some("value".to_string()),
+        min_version: Some("1.0.0".to_string()),
+        version: Some("2.0.0".to_string()),
+        install_command: Some("install".to_string()),
+        force: Some(true),
+    };
+    let second = InstallCrateInfo {
+        crate_name: "test".to_string(),
+        binary: "bin".to_string(),
+        test_arg: TestArg {
+            inner: vec!["--help".to_string()],
+        },
+        rustup_component_name: Some("value".to_string()),
+        min_version: Some("1.0.0".to_string()),
+        version: Some("2.0.0".to_string()),
+        install_command: Some("install".to_string()),
+        force: None,
+    };
+
+    assert!(first != second);
+}
+
+#[test]
+fn install_crate_info_eq_different_force_value() {
+    let first = InstallCrateInfo {
+        crate_name: "test".to_string(),
+        binary: "bin".to_string(),
+        test_arg: TestArg {
+            inner: vec!["--help".to_string()],
+        },
+        rustup_component_name: Some("value".to_string()),
+        min_version: Some("1.0.0".to_string()),
+        version: Some("2.0.0".to_string()),
+        install_command: Some("install".to_string()),
+        force: Some(true),
+    };
+    let second = InstallCrateInfo {
+        crate_name: "test".to_string(),
+        binary: "bin".to_string(),
+        test_arg: TestArg {
+            inner: vec!["--help".to_string()],
+        },
+        rustup_component_name: Some("value".to_string()),
+        min_version: Some("1.0.0".to_string()),
+        version: Some("2.0.0".to_string()),
+        install_command: Some("install".to_string()),
+        force: Some(false),
     };
 
     assert!(first != second);
@@ -1053,6 +1175,7 @@ fn install_crate_eq_same_crate_info() {
         min_version: Some("1.0.0".to_string()),
         version: Some("2.0.0".to_string()),
         install_command: Some("install".to_string()),
+        force: Some(true),
     };
     let first = InstallCrate::CrateInfo(info.clone());
     let second = InstallCrate::CrateInfo(info.clone());
@@ -1121,6 +1244,7 @@ fn install_crate_eq_different_crate_info() {
         min_version: None,
         version: None,
         install_command: Some("install".to_string()),
+        force: None,
     });
     let second = InstallCrate::CrateInfo(InstallCrateInfo {
         crate_name: "test2".to_string(),
@@ -1132,6 +1256,7 @@ fn install_crate_eq_different_crate_info() {
         min_version: None,
         version: None,
         install_command: Some("install".to_string()),
+        force: None,
     });
 
     assert!(first != second);